QuakeAlert.py
The code works, but it's probably a little messy... I'll rewrite it at some point when I have time to make it tidier, and make it easier to set thresholds for alerting.  Additionally I'll build in a feature to take into account the quake's location relative to the user - this only alerts when large earthquakes of mag 6.0 or greater (or MMI => 6) occur; local earthquakes may be smaller, but have a much greater effect.